requirements.txt
streamlit
openai
python-dotenv
joblib
langfuse